refactor

+++ nlp/textrank/textrank.h
@@ -11,8 +11,8 @@ namespace nlp
 	public:
 		text_ranker();
 		void load_stop_words(const std::vector<std::string>& stop_words_pathes);
-		bool key_sentences(const std::wstring& texts, std::vector< std::pair< std::wstring, double> >& key_sentences, int topK);
-		bool key_words(const std::wstring& texts, std::vector< std::pair< std::wstring, double> >& keywords, int topK);
+		bool key_sentences(const std::wstring& texts, std::vector< std::pair< std::wstring, double> >& key_sentences, size_t topK);
+		bool key_words(const std::wstring& texts, std::vector< std::pair< std::wstring, double> >& keywords, size_t topK);
 	private:
 		std::unique_ptr<stop_words_t> stop_words;
 		std::vector<std::wstring> n_grams;
